perf(e2e): remove unnecessary serial mode from read-only test files
## Serial Mode Removal (10 files)

- Remove `test.describe.configure({ mode: 'serial' })` from 10 read-only
  E2E test files that only navigate pages and assert on rendered DOM
- Cookie-based `appState` (createCookieStateFixture) sets per-context
  cookies — no shared DB state to protect between parallel workers
- Files: trading, team-schedule, league-schedule, waivers, career-leaderboards,
  season-leaderboards, contract-extension, voting-results, phase-gating-public,
  csrf-tokens
- 117 tests freed from serial execution across 4 Playwright workers

## File Splits (2 files → 4 files)

- Split voting.spec.ts: read-only rendering (9 tests, parallel) vs
  submission/validation (7 tests, serial in voting-submission.spec.ts)
- Split free-agency.spec.ts: read-only rendering + validation errors
  (25 tests, parallel) vs offer submission (14 tests, serial in
  free-agency-submission.spec.ts)
- Total: 151 tests freed from unnecessary serial execution

## CI Sharding

- Increase E2E shards from 3 to 4 in e2e-tests.yml to better distribute
  remaining serial files across runners
